+++ public/res/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Version 2.0.30200413 (dev)
+- Assigned "Select 1st to 8th file" to Ctrl+1~8.
 - Improved Select file.
 
 ## Version 2.0.30200412 (dev)